guardrails: lite template (lint + pip-audit + bandit + review); baseline to zero (#10)
* guardrails: lite template (lint + pip-audit + bandit + review); lint baseline to zero

Real defects the gate found on first contact:
- notebooks/06: numpy used as np in the first cell but never imported (F821)
- 13 file handles opened without a context manager, incl. the zip index
  writer that was only closed on the happy path

BLE001 (blind except) is declared in ruff.toml with its reason rather
than fixed: these scripts log-and-continue over 40k zip members by design
and there is no test suite to verify a narrower set against.

No functional change intended; ruff/pre-commit auto-fixes and formatting
otherwise.

* chore: drop a stray root-level copy of scripts/guardrails-selftest.sh

+++ analysis/05_clade_montage.py
@@ -10,14 +10,16 @@ A species whose mask cannot supply such a component is EXCLUDED and reported.
 """
 
 import os
-import sys
 import signal
+import sys
+
 import numpy as np
 from PIL import Image, ImageDraw, ImageFont
 from skimage import measure
 
 sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-from _paths import data, fig as fig_path  # noqa: E402
+from _paths import data
+from _paths import fig as fig_path
 
 signal.signal(
     signal.SIGALRM, lambda *_: (sys.stderr.write("walltime guard\n"), sys.exit(2))
@@ -92,7 +94,7 @@ for fn, (sp, vouch) in sorted(SPECIES.items(), key=lambda kv: kv[1][0]):
         continue
     comp = p.image
     h, w = comp.shape
-    new_w = max(8, int(round(w * (LEAF_H / h))))
+    new_w = max(8, round(w * (LEAF_H / h)))
     leaf = Image.fromarray(np.where(comp, 0, 255).astype(np.uint8)).resize(
         (new_w, LEAF_H), Image.LANCZOS
     )
